+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * Fixed issue in `get_zero_crossings_array_indices` where it would fail if array did not contain any zeros.
 * Added calculation of equivalent number of cycles and equivalent uniform amplitude using power law relationship as intensity measures
+* Added function `get_n_cyc_array` to compute number of cycles series from a loading series
 
 
 1.1.0 (2019-10-08)